backend/requirements.txt
anyio==3.6.2
autopep8==2.0.4 
certifi==2023.11.17 
cffi==1.15.1
charset-normalizer==3.1.0
click==8.1.3
colorama==0.4.6
cryptography==40.0.1
fastapi==0.95.0
greenlet==2.0.2
h11==0.14.0
idna==3.4
pycodestyle
pycparser==2.21
pydantic==1.10.7
PyMySQL==1.0.3
python-dotenv==1.0.0
requests==2.28.2
sniffio==1.3.0
SQLAlchemy==2.0.9
starlette==0.26.1
toml==0.10.2 
typing_extensions==4.5.0
urllib3==1.26.15
uvicorn==0.21.1
wincertstore==0.2
